+++ run_linear_driver_experiments.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+python3 -m src.linear.experiments_driver
