--- run_linear_simple_simulations.sh
@@ -1,6 +1,0 @@
-#!/bin/sh
-
-SIM_STEPS=10000
-NUM_EXPERIMENTS=5
-
-python3 -m src.linear.simulate_active_learning --num-experiments $NUM_EXPERIMENTS --simulation-steps $SIM_STEPS
